FIX-6 - В pom изменены версии для hibernate-search-orm,swagger, поправлена "битая" ссылка на репозиторий - Добавлен конфиг для swagger - Поправлены ссылки на dto для метода search в CompanyController и дописана спецификация swagger для этого метода - Request и Response перенесены в отдельные пакеты
