+++ src/main/java/com/fedag/internship/controller/CompanyController.java
@@ -1,8 +1,8 @@
 package com.fedag.internship.controller;
 
-import com.fedag.internship.domain.dto.CompanyRequest;
-import com.fedag.internship.domain.dto.CompanyRequestUpdate;
-import com.fedag.internship.domain.dto.CompanyResponse;
+import com.fedag.internship.domain.dto.request.CompanyRequest;
+import com.fedag.internship.domain.dto.request.CompanyRequestUpdate;
+import com.fedag.internship.domain.dto.response.CompanyResponse;
 import com.fedag.internship.domain.dto.DtoErrorInfo;
 import com.fedag.internship.domain.mapper.CompanyMapper;
 import com.fedag.internship.service.CompanyService;
@@ -127,10 +127,17 @@ public class CompanyController {
         return new ResponseEntity<>(OK);
     }
 
+    @Operation(summary = "Получение страницы с компаниями по критериям")
+    @ApiResponse(responseCode = "200", description = "Компании найдены",
+            content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                    schema = @Schema(implementation = Page.class))})
+    @ApiResponse(responseCode = "400", description = "Внутренняя ошибка сервера",
+            content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                    schema = @Schema(implementation = DtoErrorInfo.class))})
     @GetMapping("/search")
     public ResponseEntity<Page<CompanyResponse>> search(@RequestParam String keyword, Pageable pageable) {
         Page<CompanyResponse> companies = companyService.searchCompanyByName(keyword, pageable)
                 .map(companyMapper::toResponse);
-        return new ResponseEntity<>(companies, HttpStatus.OK);
+        return new ResponseEntity<>(companies, OK);
     }
 }
