+++ src/main/java/com/fedag/internship/controller/UserController.java
@@ -1,9 +1,9 @@
 package com.fedag.internship.controller;
 
 import com.fedag.internship.domain.dto.DtoErrorInfo;
-import com.fedag.internship.domain.dto.UserRequest;
-import com.fedag.internship.domain.dto.UserRequestUpdate;
-import com.fedag.internship.domain.dto.UserResponse;
+import com.fedag.internship.domain.dto.request.UserRequest;
+import com.fedag.internship.domain.dto.request.UserRequestUpdate;
+import com.fedag.internship.domain.dto.response.UserResponse;
 import com.fedag.internship.domain.mapper.UserMapper;
 import com.fedag.internship.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
