+++ src/main/java/com/fedag/internship/domain/dto/request/CommentRequest.java
@@ -1,5 +1,6 @@
-package com.fedag.internship.domain.dto;
+package com.fedag.internship.domain.dto.request;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,7 +10,7 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
 /**
- * class CommentUpdateDto for update Dto layer of Comment.
+ * class CommentCreateDto for create Dto layer of Comment.
  *
  * @author damir.iusupov
  * @since 2022-06-01
@@ -17,12 +18,13 @@ import javax.validation.constraints.Size;
 @Getter
 @Setter
 @NoArgsConstructor
-@Schema(name = "Обновление комментария", description = "Запрос для обновления комментария")
-public class CommentRequestUpdate {
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Schema(name = "Создание комментария", description = "Запрос для создания комментария")
+public class CommentRequest {
     @Schema(description = "Текст комментария",
             maxLength = 500,
             minLength = 1,
-            example = "some text upd")
+            example = "some text")
     @NotBlank
     @Size(max = 500)
     private String text;
