+++ src/main/java/com/fedag/internship/domain/dto/request/CommentRequestUpdate.java
@@ -1,6 +1,5 @@
-package com.fedag.internship.domain.dto;
+package com.fedag.internship.domain.dto.request;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +9,7 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
 /**
- * class CommentCreateDto for create Dto layer of Comment.
+ * class CommentUpdateDto for update Dto layer of Comment.
  *
  * @author damir.iusupov
  * @since 2022-06-01
@@ -18,13 +17,12 @@ import javax.validation.constraints.Size;
 @Getter
 @Setter
 @NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Schema(name = "Создание комментария", description = "Запрос для создания комментария")
-public class CommentRequest {
+@Schema(name = "Обновление комментария", description = "Запрос для обновления комментария")
+public class CommentRequestUpdate {
     @Schema(description = "Текст комментария",
             maxLength = 500,
             minLength = 1,
-            example = "some text")
+            example = "some text upd")
     @NotBlank
     @Size(max = 500)
     private String text;
