+++ src/main/java/com/fedag/internship/domain/dto/request/UserRequest.java
@@ -1,4 +1,4 @@
-package com.fedag.internship.domain.dto;
+package com.fedag.internship.domain.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import io.swagger.v3.oas.annotations.media.Schema;
