+++ src/main/java/com/fedag/internship/domain/dto/response/CommentResponse.java
@@ -1,4 +1,4 @@
-package com.fedag.internship.domain.dto;
+package com.fedag.internship.domain.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
