+++ src/main/java/com/fedag/internship/domain/dto/response/CompanyResponse.java
@@ -1,52 +1,51 @@
-package com.fedag.internship.domain.dto;
+package com.fedag.internship.domain.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.Size;
+import lombok.experimental.Accessors;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@Schema(name = "Обновление компании", description = "Запрос для обновления компании")
-public class CompanyRequestUpdate {
+@Accessors(chain = true)
+@Schema(name = "Компания", description = "Информация о компании")
+public class CompanyResponse {
+    @Schema(description = "Идентификатор компании",
+            example = "1")
+    private Long id;
+
     @Schema(description = "Имя компании",
             maxLength = 255,
             minLength = 1,
-            example = "some name upd")
-    @Size(max = 255)
+            example = "some name")
     private String name;
 
     @Schema(description = "Описание компании",
             maxLength = 500,
             minLength = 1,
-            example = "some description upd")
-    @Size(max = 500)
+            example = "some description")
     private String description;
 
     @Schema(description = "Рейтинг компании",
             maximum = "10",
             minimum = "1",
-            example = "4")
-    @Min(1)
-    @Max(10)
+            example = "5")
     private Double rating;
 
     @Schema(description = "Место нахождения компании компании",
             maxLength = 255,
             minLength = 1,
-            example = "some location upd")
-    @Size(max = 255)
+            example = "some location")
     private String location;
 
     @Schema(description = "Ссылка на компанию",
             maxLength = 255,
             minLength = 1,
-            example = "someUrlUpd.com")
-    @Size(max = 255)
+            example = "someUrl.com")
     private String link;
+
+    @Schema(description = "Идентификатор пользователя компании", example = "1")
+    private Long userId;
 }
