+++ src/main/java/com/fedag/internship/domain/mapper/CompanyMapper.java
@@ -1,8 +1,8 @@
 package com.fedag.internship.domain.mapper;
 
-import com.fedag.internship.domain.dto.CompanyRequest;
-import com.fedag.internship.domain.dto.CompanyRequestUpdate;
-import com.fedag.internship.domain.dto.CompanyResponse;
+import com.fedag.internship.domain.dto.request.CompanyRequest;
+import com.fedag.internship.domain.dto.request.CompanyRequestUpdate;
+import com.fedag.internship.domain.dto.response.CompanyResponse;
 import com.fedag.internship.domain.entity.CompanyEntity;
 
 public interface CompanyMapper {
