+++ src/main/java/com/fedag/internship/domain/mapper/impl/UserMapperImpl.java
@@ -1,9 +1,9 @@
 package com.fedag.internship.domain.mapper.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fedag.internship.domain.dto.UserRequest;
-import com.fedag.internship.domain.dto.UserRequestUpdate;
-import com.fedag.internship.domain.dto.UserResponse;
+import com.fedag.internship.domain.dto.request.UserRequest;
+import com.fedag.internship.domain.dto.request.UserRequestUpdate;
+import com.fedag.internship.domain.dto.response.UserResponse;
 import com.fedag.internship.domain.entity.CommentEntity;
 import com.fedag.internship.domain.entity.UserEntity;
 import com.fedag.internship.domain.mapper.UserMapper;
