+++ src/main/java/com/fedag/internship/service/CompanyService.java
@@ -16,6 +16,5 @@ public interface CompanyService {
 
     void deleteCompany(Long id);
 
-    Page<CompanyDto> searchCompanyByName(String keyword, Pageable pageable);
-
+    Page<CompanyEntity> searchCompanyByName(String keyword, Pageable pageable);
 }
